requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
transformers==4.46.0
torch
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
Levenshtein==0.23.0 
fasttext-wheel==0.9.2
openai>=1.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rapidfuzz==3.6.1
numpy==1.26.4
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
pytest-asyncio>=1.0.0
dotenv